Fie26

+++ app/src/main/java/com/fieldbook/tracker/traitLayouts/PhotoTraitLayout.java
@@ -1,24 +1,23 @@
 package com.fieldbook.tracker.traitLayouts;
 
 import android.app.Activity;
+import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.graphics.Color;
 import android.graphics.Matrix;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.media.ExifInterface;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Handler;
-import android.os.SystemClock;
 import android.provider.MediaStore;
-import android.text.TextWatcher;
+import android.text.Html;
 import android.util.AttributeSet;
 import android.util.Log;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.EditText;
@@ -43,7 +42,6 @@ import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
@@ -76,11 +74,17 @@ public class PhotoTraitLayout extends TraitLayout {
 
     @Override
     public void loadLayout(){
-
         getEtCurVal().removeTextChangedListener(getCvText());
         getEtCurVal().removeTextChangedListener(getCvNum());
         getEtCurVal().setVisibility(EditText.GONE);
         getEtCurVal().setEnabled(false);
+
+        // Run saving task in the background so we can showing progress dialog
+        Handler mHandler = new Handler();
+        mHandler.post(importRunnable);
+    }
+
+    public void loadLayoutWork() {
 
         // Always set to null as default, then fill in with trait value
         photoLocation = new ArrayList<>();
@@ -95,22 +99,6 @@ public class PhotoTraitLayout extends TraitLayout {
             for (int i = 0; i < photoLocation.size(); i++) {
                 drawables.add(new BitmapDrawable(displayScaledSavedPhoto(photoLocation.get(i))));
             }
-
-            photoAdapter = new GalleryImageAdapter((Activity) getContext(), drawables);
-            photo.setAdapter(photoAdapter);
-            photo.setSelection(photo.getCount() - 1);
-            photo.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-
-                @Override
-                public void onItemClick(AdapterView<?> arg0,
-                                        View arg1, int pos, long arg3) {
-                    displayPlotImage(photoLocation.get(photo.getSelectedItemPosition()));
-                }
-            });
-
-        } else {
-            photoAdapter = new GalleryImageAdapter((Activity) getContext(), drawables);
-            photo.setAdapter(photoAdapter);
         }
 
         if (!getNewTraits().containsKey(getCurrentTrait().getTrait())) {
@@ -119,9 +107,70 @@ public class PhotoTraitLayout extends TraitLayout {
             }
         }
     }
+
+    // Creates a new thread to do importing
+    private Runnable importRunnable = new Runnable() {
+        public void run() {
+            new PhotoTraitLayout.LoadImagesRunnableTask().execute(0);
+        }
+    };
+
+    // Mimics the class used in the csv field importer to run the saving
+    // task in a different thread from the UI thread so the app doesn't freeze up.
+    private class LoadImagesRunnableTask extends AsyncTask<Integer, Integer, Integer> {
+
+        ProgressDialog dialog;
+
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+            dialog = new ProgressDialog(getContext());
+            dialog.setIndeterminate(true);
+            dialog.setCancelable(false);
+            dialog.setMessage(Html.fromHtml(getContext().getResources().getString(R.string.images_loading)));
+            dialog.show();
+        }
+
+        @Override
+        protected Integer doInBackground(Integer... params) {
+            loadLayoutWork();
+            return 0;
+        }
+
+        @Override
+        protected void onPostExecute(Integer result) {
+            if (dialog.isShowing())
+                dialog.dismiss();
+
+            File img = new File(Constants.PLOTDATAPATH + "/" + getPrefs().getString("FieldFile", "") + "/" + "/photos/");
+            if (img.listFiles() != null) {
+
+                photoAdapter = new GalleryImageAdapter((Activity) getContext(), drawables);
+                photo.setAdapter(photoAdapter);
+                photo.setSelection(photo.getCount() - 1);
+                photo.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+
+                    @Override
+                    public void onItemClick(AdapterView<?> arg0,
+                                            View arg1, int pos, long arg3) {
+                        displayPlotImage(photoLocation.get(photo.getSelectedItemPosition()));
+                    }
+                });
+
+            } else {
+                photoAdapter = new GalleryImageAdapter((Activity) getContext(), drawables);
+                photo.setAdapter(photoAdapter);
+            }
+        }
+    }
+
     @Override
     public void deleteTraitListener() {
-        deletePhotoWarning();
+        deletePhotoWarning(false, null);
+    }
+
+    public void brapiDelete(Map newTraits) {
+        deletePhotoWarning(true, newTraits);
     }
 
     private Bitmap displayScaledSavedPhoto(String path) {
@@ -231,7 +280,7 @@ public class PhotoTraitLayout extends TraitLayout {
         mediaScanIntent.setData(contentUri);
         getContext().sendBroadcast(mediaScanIntent);
 
-        updateTraitAllowDuplicates(currentTrait.getTrait(), "photo", file.getAbsolutePath(), newTraits);
+        updateTraitAllowDuplicates(currentTrait.getTrait(), "photo", file.getAbsolutePath(), null, newTraits);
 
         photoAdapter = new GalleryImageAdapter((Activity) getContext(), drawables);
 
@@ -239,28 +288,32 @@ public class PhotoTraitLayout extends TraitLayout {
         photo.setSelection(photoAdapter.getCount() - 1);
     }
 
-    private void updateTraitAllowDuplicates(String parent, String trait, String value, Map newTraits) {
+    private void updateTraitAllowDuplicates(String parent, String trait, String value, String newValue, Map newTraits) {
 
-        if (getCRange() == null || getCRange().plot_id.length() == 0) {
-            return;
+        if (!value.equals(newValue)) {
+
+            if (getCRange() == null || getCRange().plot_id.length() == 0) {
+                return;
+            }
+
+            Log.d("Field Book", trait + " " + value);
+
+            if (newTraits.containsKey(parent))
+                newTraits.remove(parent);
+
+            newTraits.put(parent, value);
+
+            //Observation observation = ConfigActivity.dt.getObservation(getCRange().plot_id, parent);
+            Observation observation = ConfigActivity.dt.getObservationByValue(getCRange().plot_id, parent, value);
+
+            ConfigActivity.dt.deleteTraitByValue(getCRange().plot_id, parent, value);
+
+            String exp_id = Integer.toString(getPrefs().getInt("ExpID", 0));
+            ConfigActivity.dt.insertUserTraits(getCRange().plot_id, parent, trait, newValue == null ? value : newValue, getPrefs().getString("FirstName", "") + " " + getPrefs().getString("LastName", ""), getPrefs().getString("Location", ""), "", exp_id, observation.getDbId(), observation.getLastSyncedTime()); //TODO add notes and exp_id
         }
-
-        Log.d("Field Book",trait + " " + value);
-
-        if (newTraits.containsKey(parent))
-            newTraits.remove(parent);
-
-        newTraits.put(parent, value);
-
-        Observation observation = ConfigActivity.dt.getObservation(getCRange().plot_id, parent);
-
-        ConfigActivity.dt.deleteTraitByValue(getCRange().plot_id, parent, value);
-
-        String exp_id = Integer.toString(getPrefs().getInt("ExpID", 0));
-        ConfigActivity.dt.insertUserTraits(getCRange().plot_id, parent, trait, value, getPrefs().getString("FirstName","") + " " + getPrefs().getString("LastName",""), getPrefs().getString("Location",""),"",exp_id, observation.getDbId(), observation.getLastSyncedTime()); //TODO add notes and exp_id
     }
 
-    private void deletePhotoWarning() {
+    private void deletePhotoWarning(final Boolean brapiDelete, final Map newTraits) {
         AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
 
         builder.setTitle(getContext().getString(R.string.dialog_warning));
@@ -271,26 +324,45 @@ public class PhotoTraitLayout extends TraitLayout {
             public void onClick(DialogInterface dialog, int which) {
                 dialog.dismiss();
 
+                if (brapiDelete) {
+                    Toast.makeText(getContext().getApplicationContext(), getContext().getString(R.string.brapi_delete_message), Toast.LENGTH_SHORT).show();
+                    //updateTrait(parent, currentTrait.getFormat(), getString(R.string.brapi_na));
+                }
+
                 if (photo.getCount() > 0) {
                     String item = photoLocation.get(photo.getSelectedItemPosition());
-                    photoLocation.remove(photo.getSelectedItemPosition());
-                    drawables.remove(photo.getSelectedItemPosition());
+                    if (!brapiDelete) {
+                        photoLocation.remove(photo.getSelectedItemPosition());
+                        drawables.remove(photo.getSelectedItemPosition());
+                    }
 
                     File f = new File(item);
                     f.delete();
                     Utils.scanFile((Activity) getContext(),f);
 
                     // Remove individual images
-                    ConfigActivity.dt.deleteTraitByValue(getCRange().plot_id, getCurrentTrait().getTrait(), item);
+                    if (brapiDelete) {
+                        updateTraitAllowDuplicates(getCurrentTrait().getTrait(), "photo", item, "NA", newTraits);
+                        //ConfigActivity.dt.updateTraitByValue(getCRange().plot_id, getCurrentTrait().getTrait(), item, "NA");
+                        loadLayout();
+                    }
+                    else {
+                        ConfigActivity.dt.deleteTraitByValue(getCRange().plot_id, getCurrentTrait().getTrait(), item);
+                    }
 
                     // Only do a purge by trait when there are no more images left
-                    if (photoLocation.size() == 0)
-                        removeTrait(getCurrentTrait().getTrait());
+                    if (!brapiDelete) {
+                        if (photoLocation.size() == 0)
+                            removeTrait(getCurrentTrait().getTrait());
+                    }
 
                     photoAdapter = new GalleryImageAdapter((Activity) getContext(), drawables);
 
                     photo.setAdapter(photoAdapter);
-                } else {
+                }
+                else {
+                    // If an NA exists, delete it
+                    ConfigActivity.dt.deleteTraitByValue(getCRange().plot_id, getCurrentTrait().getTrait(), "NA");
                     ArrayList<Drawable> emptyList = new ArrayList<>();
 
                     photoAdapter = new GalleryImageAdapter((Activity) getContext(), emptyList);
